Refactor
Put CreateNavS into NavigationStructure instead of Models.
