+++ zssmodel/src/org/zkoss/zss/model/impl/Hybrid_Model.java
@@ -796,6 +796,7 @@ public class Hybrid_Model extends RCV_Model {
 
             this.navSbuckets  = model.navSbuckets;
             this.navS = model.navS;
+            this.navS.setCurrentSheet(sheet);
             //this.indexString = model.indexString;
 
             CellRegion range = model.getBounds(dbContext);
